fix: Accounts Important Issues 修正
変更内容:
- Contract重複解消: IdContract基底クラス作成、Show/Destroyが継承
- ログ出力追加: Create/Update/Destroy に操作ログ追加

変更理由:
- DRY原則準拠（Contract重複解消）
- 監査ログ対応（誰が何をいつ操作したか追跡可能に）

影響範囲:
- app/contracts/accounts/id_contract.rb（新規）
- app/contracts/accounts/show.rb, destroy.rb
- app/services/accounts/create.rb, update.rb, destroy.rb

テスト結果: Pass (1047 examples, 0 failures)
