+++ app/contracts/accounts/show.rb
@@ -2,30 +2,10 @@
 
 module Contracts
   module Accounts
-    class Show
-      include ActiveModel::Model
-
-      attr_accessor :id
-
-      validates :id, presence: true, numericality: { only_integer: true }
-
-      def self.call(params)
-        contract = new(
-          id: params[:id]
-        )
-
-        if contract.valid?
-          Result.new(success?: true, value: contract.normalized_attributes, errors: [])
-        else
-          Result.new(success?: false, value: nil, errors: contract.errors.full_messages)
-        end
-      end
-
-      def normalized_attributes
-        {
-          id:
-        }
-      end
+    # アカウント詳細取得用Contract
+    #
+    # IDのみを検証（IdContractを継承）
+    class Show < IdContract
     end
   end
 end
